Make offset type specific to avoid errors on non-linux systems. (#84)

+++ src/relax/vm/memory_manager.cc
@@ -76,7 +76,7 @@ inline size_t GetDataAlignment(const DLTensor& arr) {
   return align;
 }
 
-runtime::NDArray StorageObj::AllocNDArray(size_t offset, ShapeTuple shape, DLDataType dtype) {
+runtime::NDArray StorageObj::AllocNDArray(uint64_t offset, ShapeTuple shape, DLDataType dtype) {
   VerifyDataType(dtype);
 
   // critical zone: allocate header, cannot throw
